+++ experimental/transport-otlp-http/src/payload/transform/types.ts
@@ -1,9 +1,12 @@
-import type { IKeyValue, IResource } from '@opentelemetry/otlp-transformer/build/src/common/internal-types';
+import type {
+  IKeyValue,
+  Resource as OtelResource,
+} from '@opentelemetry/otlp-transformer/build/src/common/internal-types';
 import type { IResourceSpans } from '@opentelemetry/otlp-transformer/build/src/trace/internal-types';
 
 import type { APIEvent, Meta, TraceEvent, TransportItem } from '@grafana/faro-core';
 
-export interface Resource extends Partial<Pick<IResource, 'droppedAttributesCount'>> {
+export interface Resource extends Partial<Pick<OtelResource, 'droppedAttributesCount'>> {
   attributes: IKeyValue[];
 }