+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next
 
 - Improvement (`@grafana/faro-web-sdk`): Provide a function to start a user-action (#1115).
+- Dependencies (`@grafana/faro-*`): Upgrade OpenTelemetry dependencies to v0.200.\* (#1126).
 
 ## 1.15.0